fixed erroneous git issue with filename change
